@@ -4,7 +4,9 @@ import (
 	"net/url"
 	"testing"
 
+	"github.com/projectdiscovery/nuclei/v2/internal/severity"
 	"github.com/projectdiscovery/nuclei/v2/internal/testutils"
+	"github.com/projectdiscovery/nuclei/v2/pkg/model"
 	"github.com/stretchr/testify/require"
 )
 
@@ -16,6 +18,45 @@ func TestBaseURLWithTemplatePrefs(t *testing.T) {
 	data, parsed = baseURLWithTemplatePrefs(data, parsed)
 	require.Equal(t, "http://localhost:8000/test", parsed.String(), "could not get correct value")
 	require.Equal(t, "{{BaseURL}}/newpath", data, "could not get correct data")
+}
+
+func TestVariables(t *testing.T) {
+	baseURL := "http://localhost:9001/test/123"
+	parsed, _ := url.Parse(baseURL)
+	values := generateVariables(parsed, true)
+
+	require.Equal(t, values["BaseURL"], parsed.String(), "incorrect baseurl")
+	require.Equal(t, values["RootURL"], "http://localhost:9001", "incorrect rootURL")
+	require.Equal(t, values["Host"], "localhost", "incorrect domain name")
+	require.Equal(t, values["Path"], "/test", "incorrect path")
+	require.Equal(t, values["File"], "123", "incorrect file")
+	require.Equal(t, values["Port"], "9001", "incorrect port number")
+	require.Equal(t, values["Scheme"], "http", "incorrect scheme")
+	require.Equal(t, values["Hostname"], "localhost:9001", "incorrect hostname")
+
+	baseURL = "https://example.com"
+	parsed, _ = url.Parse(baseURL)
+	values = generateVariables(parsed, false)
+
+	require.Equal(t, values["BaseURL"], parsed.String(), "incorrect baseurl")
+	require.Equal(t, values["Host"], "example.com", "incorrect domain name")
+	require.Equal(t, values["RootURL"], "https://example.com", "incorrect rootURL")
+	require.Equal(t, values["Path"], "", "incorrect path")
+	require.Equal(t, values["Port"], "443", "incorrect port number")
+	require.Equal(t, values["Scheme"], "https", "incorrect scheme")
+	require.Equal(t, values["Hostname"], "example.com", "incorrect hostname")
+
+	baseURL = "ftp://foobar.com/"
+	parsed, _ = url.Parse(baseURL)
+	values = generateVariables(parsed, true)
+
+	require.Equal(t, values["BaseURL"], parsed.String(), "incorrect baseurl")
+	require.Equal(t, values["Host"], "foobar.com", "incorrect domain name")
+	require.Equal(t, values["RootURL"], "ftp://foobar.com", "incorrect rootURL")
+	require.Equal(t, values["Path"], "", "incorrect path")
+	require.Equal(t, values["Port"], "", "incorrect port number") // Unsupported protocol results in a blank port
+	require.Equal(t, values["Scheme"], "ftp", "incorrect scheme")
+	require.Equal(t, values["Hostname"], "foobar.com", "incorrect hostname")
 }
 
 func TestMakeRequestFromModal(t *testing.T) {
@@ -36,7 +77,7 @@ func TestMakeRequestFromModal(t *testing.T) {
 	}
 	executerOpts := testutils.NewMockExecuterOptions(options, &testutils.TemplateInfo{
 		ID:   templateID,
-		Info: map[string]interface{}{"severity": "low", "name": "test"},
+		Info: model.Info{SeverityHolder: severity.SeverityHolder{Severity: severity.Low}, Name: "test"},
 	})
 	err := request.Compile(executerOpts)
 	require.Nil(t, err, "could not compile http request")
@@ -63,7 +104,7 @@ func TestMakeRequestFromModalTrimSuffixSlash(t *testing.T) {
 	}
 	executerOpts := testutils.NewMockExecuterOptions(options, &testutils.TemplateInfo{
 		ID:   templateID,
-		Info: map[string]interface{}{"severity": "low", "name": "test"},
+		Info: model.Info{SeverityHolder: severity.SeverityHolder{Severity: severity.Low}, Name: "test"},
 	})
 	err := request.Compile(executerOpts)
 	require.Nil(t, err, "could not compile http request")
@@ -101,7 +142,7 @@ Accept-Encoding: gzip`},
 	}
 	executerOpts := testutils.NewMockExecuterOptions(options, &testutils.TemplateInfo{
 		ID:   templateID,
-		Info: map[string]interface{}{"severity": "low", "name": "test"},
+		Info: model.Info{SeverityHolder: severity.SeverityHolder{Severity: severity.Low}, Name: "test"},
 	})
 	err := request.Compile(executerOpts)
 	require.Nil(t, err, "could not compile http request")
@@ -140,7 +181,7 @@ Accept-Encoding: gzip`},
 	}
 	executerOpts := testutils.NewMockExecuterOptions(options, &testutils.TemplateInfo{
 		ID:   templateID,
-		Info: map[string]interface{}{"severity": "low", "name": "test"},
+		Info: model.Info{SeverityHolder: severity.SeverityHolder{Severity: severity.Low}, Name: "test"},
 	})
 	err := request.Compile(executerOpts)
 	require.Nil(t, err, "could not compile http request")
