@@ -6,13 +6,14 @@ import (
 	"io/ioutil"
 	"strings"
 
-	jira "github.com/andygrunwald/go-jira"
+	"github.com/andygrunwald/go-jira"
+	"github.com/projectdiscovery/gologger"
 	"github.com/projectdiscovery/nuclei/v2/pkg/output"
 	"github.com/projectdiscovery/nuclei/v2/pkg/reporting/format"
 	"github.com/projectdiscovery/nuclei/v2/pkg/types"
 )
 
-// Integration is a client for a issue tracker integration
+// Integration is a client for an issue tracker integration
 type Integration struct {
 	jira    *jira.Client
 	options *Options
@@ -22,6 +23,8 @@ type Integration struct {
 type Options struct {
 	// Cloud value is set to true when Jira cloud is used
 	Cloud bool `yaml:"cloud"`
+	// UpdateExisting value if true, the existing opened issue is updated
+	UpdateExisting bool `yaml:"update-existing"`
 	// URL is the URL of the jira server
 	URL string `yaml:"url"`
 	// AccountID is the accountID of the jira user.
@@ -53,8 +56,8 @@ func New(options *Options) (*Integration, error) {
 	return &Integration{jira: jiraClient, options: options}, nil
 }
 
-// CreateIssue creates an issue in the tracker
-func (i *Integration) CreateIssue(event *output.ResultEvent) error {
+// CreateNewIssue creates a new issue in the tracker
+func (i *Integration) CreateNewIssue(event *output.ResultEvent) error {
 	summary := format.Summary(event)
 
 	fields := &jira.IssueFields{
@@ -91,33 +94,82 @@ func (i *Integration) CreateIssue(event *output.ResultEvent) error {
 	return nil
 }
 
+// CreateIssue creates an issue in the tracker or updates the existing one
+func (i *Integration) CreateIssue(event *output.ResultEvent) error {
+	if i.options.UpdateExisting {
+		issueID, err := i.FindExistingIssue(event)
+		if err != nil {
+			return err
+		} else if issueID != "" {
+			_, _, err = i.jira.Issue.AddComment(issueID, &jira.Comment{
+				Body: jiraFormatDescription(event),
+			})
+			return err
+		}
+	}
+	return i.CreateNewIssue(event)
+}
+
+// FindExistingIssue checks if the issue already exists and returns its ID
+func (i *Integration) FindExistingIssue(event *output.ResultEvent) (string, error) {
+	template := format.GetMatchedTemplate(event)
+	jql := fmt.Sprintf("summary ~ \"%s\" AND summary ~ \"%s\" AND status = \"Open\"", template, event.Host)
+
+	searchOptions := &jira.SearchOptions{
+		MaxResults: 1, // if any issue exists, then we won't create a new one
+	}
+
+	chunk, resp, err := i.jira.Issue.Search(jql, searchOptions)
+	if err != nil {
+		var data string
+		if resp != nil && resp.Body != nil {
+			d, _ := ioutil.ReadAll(resp.Body)
+			data = string(d)
+		}
+		return "", fmt.Errorf("%s => %s", err, data)
+	}
+
+	switch resp.Total {
+	case 0:
+		return "", nil
+	case 1:
+		return chunk[0].ID, nil
+	default:
+		gologger.Warning().Msgf("Discovered multiple opened issues %s for the host %s: The issue [%s] will be updated.", template, event.Host, chunk[0].ID)
+		return chunk[0].ID, nil
+	}
+}
+
 // jiraFormatDescription formats a short description of the generated
 // event by the nuclei scanner in Jira format.
-func jiraFormatDescription(event *output.ResultEvent) string {
+func jiraFormatDescription(event *output.ResultEvent) string { // TODO remove the code duplication: format.go <-> jira.go
 	template := format.GetMatchedTemplate(event)
 
 	builder := &bytes.Buffer{}
 	builder.WriteString("*Details*: *")
 	builder.WriteString(template)
 	builder.WriteString("* ")
+
 	builder.WriteString(" matched at ")
 	builder.WriteString(event.Host)
+
 	builder.WriteString("\n\n*Protocol*: ")
 	builder.WriteString(strings.ToUpper(event.Type))
+
 	builder.WriteString("\n\n*Full URL*: ")
 	builder.WriteString(event.Matched)
+
 	builder.WriteString("\n\n*Timestamp*: ")
 	builder.WriteString(event.Timestamp.Format("Mon Jan 2 15:04:05 -0700 MST 2006"))
+
 	builder.WriteString("\n\n*Template Information*\n\n| Key | Value |\n")
-	for k, v := range event.Info {
-		if k == "reference" {
-			continue
-		}
-		builder.WriteString(fmt.Sprintf("| %s | %s |\n", k, v))
-	}
+	builder.WriteString(format.ToMarkdownTableString(&event.Info))
+
 	builder.WriteString("\n*Request*\n\n{code}\n")
 	builder.WriteString(event.Request)
-	builder.WriteString("\n{code}\n\n*Response*\n\n{code}\n")
+	builder.WriteString("\n{code}\n")
+
+	builder.WriteString("\n*Response*\n\n{code}\n")
 	// If the response is larger than 5 kb, truncate it before writing.
 	if len(event.Response) > 5*1024 {
 		builder.WriteString(event.Response[:5*1024])
@@ -174,23 +226,17 @@ func jiraFormatDescription(event *output.ResultEvent) string {
 			builder.WriteString("\n{code}\n")
 		}
 	}
-	if d, ok := event.Info["reference"]; ok {
-		builder.WriteString("\nReference: \n")
 
-		switch v := d.(type) {
-		case string:
-			if !strings.HasPrefix(v, "-") {
-				builder.WriteString("- ")
-			}
-			builder.WriteString(v)
-		case []interface{}:
-			slice := types.ToStringSlice(v)
-			for i, item := range slice {
-				builder.WriteString("- ")
-				builder.WriteString(item)
-				if len(slice)-1 != i {
-					builder.WriteString("\n")
-				}
+	reference := event.Info.Reference
+	if !reference.IsEmpty() {
+		builder.WriteString("\nReferences: \n")
+
+		referenceSlice := reference.ToSlice()
+		for i, item := range referenceSlice {
+			builder.WriteString("- ")
+			builder.WriteString(item)
+			if len(referenceSlice)-1 != i {
+				builder.WriteString("\n")
 			}
 		}
 	}
